+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Version 0.3.0
 + Add ``config.config.static_source``
 + Add ``config.config.static_output``
 + Support Non-meta post.
++ Support customize permalink for post
 + Support TXT format post.
 + Correct the url for Twitter Card Support
 + Drop file-based cache system.